LayoutTests/editing/pasteboard/paste-text-in-link-expected.txt

This test verifies that replacing all the text inside of a link by pasting does not leave the link behind with its original href.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS containsLink(editor1) is false
PASS containsLink(editor2) is false
PASS containsLink(editor3) is false
PASS successfullyParsed is true

TEST COMPLETE
